+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Unreleased]
+### Changed
+- `JSONStreamer`: added `parse()` function
+
 ## [1.3] - 2023-06-04
 ### Changed
 - `pom.xml`: updated dependency version